fix(dispatch): HEAT meter ghosts while the baseline forges; THE RACE shared with the admin board
