+++ MEMORY.md
@@ -232,6 +232,8 @@ Side Course B (Python). Backend recon phase. Next queued task: `notes/comdata-re
 
 ## BEHAVIOR LOG
 
+- **2026-08-15 · Dispatch board: HEAT meter un-vanished + THE RACE unified (fix, v1.182.2).** Jason: "the other session took away the grind meter… the agent leaderboard is not designed and using the same logic as the admins." (1) The meter wasn't removed — #428's personal mode returns hasLadder:false until 3 active booking weeks exist (Brandie's bookings only started crediting to her at #428), and GrindMeterView returned null → vanished. Ghost-rule fix: personal mode now renders the HEAT board ghosted while the baseline forges — dashed cells, "forges after 3 active booking weeks — N of 3 logged," criterion stated; Grind gains optional activeWeeks/weeksNeeded (computePersonalGrind + empty case populate them; test added). (2) Her leaderboard was the old TopAgents (revenue-ranked, comic-era) while the admin's Agents tab runs the forged quarter RACE plate (quarterStandings + contenders + REIGNING CHAMP). Extracted the plate verbatim into shared components/dashboard/agents/RaceBoard.tsx (self-contained: useAgents + quarter logic inside) — consumed by BOTH AgentsTab and DispatchDashboard (anti-drift, BandedLadder pattern). DELETED: TopAgents.tsx, getTopAgentsByRevenue + collectRecentAgents from lib/metrics/dashboard (AgentRevenue + agentGross KEPT — getAgentGrossTable uses them; first cut sliced too deep, restored from HEAD and re-cut surgically). Tests 653.
+
 - **2026-08-15 · Dispatch Forge achievements to the hardware grammar (fix, v1.182.1).** Jason: "The dispatcher forge needs to have the achievements look like the design we created." The v1.182.0 Forge page had rolled its own award visuals — emoji icon map (🪙👑🔒 + fallback), hand-rolled TIER_RING circles, gradient Bar, comic-cream #f5e6c8. Replaced with the approved hardware grammar shared with the driver's board: earned coins = forge/Coin struck metal by tier (bronze/bronze/silver/gold/platinum map), unearned = dashed ghost circle + hint (ghost rule); tags = punched amber chip (name · badge) / dashed ghost + hint; season crowns = die-stamp plate "SEASON CROWN" / dashed "UNCLAIMED"; ALL progress = MeterCells (now EXPORTED from HardwareBoard so the two surfaces cannot drift — BandedLadder lesson applied). Kept everything #428 ordered: page structure, hero, NEXT UP within-reach sorting, how-to-earn line on every award, personal-pace basis, pops. The ★ rank glyph stays (house grammar). Emoji gate on the page: only ★ remains.
 
 _Append a dated entry whenever Jason requests new behavior or functionality from me. Newest at top._
