+++ frontend/package.json
@@ -1,7 +1,7 @@
 {
   "name": "frontend",
   "private": true,
-  "version": "1.182.1",
+  "version": "1.182.2",
   "type": "module",
   "scripts": {
     "dev": "vite",
